move btag_impl up into containing module
